requirements.txt
numpy
keras
spicy
pillow
torch
tensorflow
facenet-pytorch
streamlit